Move typing module into top level scipp
