+++ python/src/scipp/__init__.py
@@ -46,7 +46,7 @@ from .extend_units import *
 from .html import to_html, make_html
 from .object_list import _repr_html_
 from .utils import collapse, slices
-from .utils.typing import is_variable, is_dataset, is_data_array, \
+from .typing import is_variable, is_dataset, is_data_array, \
                           is_dataset_or_array
 from .compat.dict import to_dict, from_dict
 from .sizes import _make_sizes
