+++ python/src/scipp/_comparison.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from ._scipp import core as _cpp
 from ._cpp_wrapper_util import call_func as _call_cpp_func
-from .utils.typing import DataArrayLike, DatasetLike
+from .typing import DataArrayLike, DatasetLike
 
 
 def less(x: DataArrayLike, y: DataArrayLike) -> DataArrayLike:
