+++ python/src/scipp/_shape.py
@@ -8,7 +8,7 @@ from typing import Dict, List, Optional, Sequence, Tuple, Union
 
 from ._scipp import core as _cpp
 from ._cpp_wrapper_util import call_func as _call_cpp_func
-from .utils.typing import DataArrayLike, DatasetLike
+from .typing import DataArrayLike, DatasetLike
 
 
 def broadcast(x: _cpp.Variable, dims: Union[List[str], Tuple[str]],
