+++ python/src/scipp/_unary.py
@@ -1,9 +1,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2021 Scipp contributors (https://github.com/scipp)
 # @author Matthew Andrew
+from typing import Union as _Union
+
 from ._scipp import core as _cpp
 from ._cpp_wrapper_util import call_func as _call_cpp_func
-from typing import Union as _Union
 
 
 def isnan(x: _cpp.Variable) -> _cpp.Variable:
