+++ python/src/scipp/_variable.py
@@ -1,12 +1,15 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2021 Scipp contributors (https://github.com/scipp)
 # @author Matthew Andrew
-from ._scipp import core as _cpp
-from ._cpp_wrapper_util import call_func as _call_cpp_func
+
 from typing import Any as _Any, Sequence as _Sequence, Union as _Union,\
     Optional as _Optional
+
 import numpy as _np
 from numpy.typing import ArrayLike as _ArrayLike
+
+from ._scipp import core as _cpp
+from ._cpp_wrapper_util import call_func as _call_cpp_func
 
 
 def _parse_dims_shape_sizes(dims, shape, sizes):
