+++ python/src/scipp/utils/__init__.py
@@ -7,7 +7,6 @@
 
 from .collapse_and_slices import *
 from .colors import *
-from .typing import *
 from .to_string import *
 from .comparison import *
 from .get import *
